+++ assets/javascripts/kitten/components/molecules/cards/backing-card/styles.js
@@ -1,7 +1,8 @@
 import styled from 'styled-components'
 import COLORS from '../../../../constants/colors-config'
 import TYPOGRAPHY from '../../../../constants/typography-config'
-import { pxToRem, stepToRem } from '../../../../helpers/utils/typography'
+import { pxToRem } from '../../../../helpers/utils/typography'
+import { mq } from '../../../../constants/screen-config'
 
 export const StyledBackingCard = styled.article`
   --backingCard--border-width: 0;
@@ -37,7 +38,7 @@ export const StyledBackingCard = styled.article`
 
   .k-BackingCard__gridWrapper {
     flex: 1 1 auto;
-    padding: ${pxToRem(20)};
+    padding: ${pxToRem(10)} ${pxToRem(20)} ${pxToRem(20)};
     display: grid;
     grid-template-columns: [row-start] var(--backingCard--grid-col) [row-end];
     grid-gap: ${pxToRem(20)};
@@ -101,32 +102,57 @@ export const StyledBackingCard = styled.article`
     height: ${pxToRem(44)};
     display: flex;
     align-items: center;
-    justify-content: center;
+  }
+
+  .k-BackingCard__title {
+    width: 100%;
+    white-space: nowrap;
+  }
+
+  .k-BackingCard__amount__wrapper {
+    margin: 0;
   }
 
   .k-BackingCard__amount {
     position: relative;
     top: ${pxToRem(-5)};
-    font-size: ${stepToRem(5)};
+    font-size: ${pxToRem(28)};
     color: var(--color-primary-500);
+    line-height: 0.9em;
+    letter-spacing: -0.015em !important;
+  }
+
+  .k-BackingCard__amount__suffix {
+    font-size: ${pxToRem(10)};
+    line-height: ${pxToRem(22)};
+    color: var(--color-grey-700);
   }
 
   .k-BackingCard__descriptionWrapper {
-    overflow: hidden;
-    position: relative;
+    --backingCard-description-fontSize: ${pxToRem(12)};
 
-    &.k-BackingCard__descriptionWrapper--truncateText {
-      max-height: calc(3 * 1.5 * ${stepToRem(-2)});
+    @media ${mq.tabletAndDesktop} {
+      --backingCard-description-fontSize: ${pxToRem(14)};
+    }
 
-      .k-BackingCard__description__moreButton {
-        line-height: 1.5;
-      }
+    .k-BackingCard__description--truncateText {
+      overflow: hidden;
+      position: relative;
+      max-height: calc(
+        6 * (var(--backingCard-description-fontSize) + ${pxToRem(4)})
+      );
+    }
+
+    .k-BackingCard__description__moreButton {
+      margin-top: ${pxToRem(10)};
     }
 
     .k-BackingCard__description * {
       ${TYPOGRAPHY.fontStyles.light.fontFamily};
-      font-size: ${stepToRem(-2)};
-      line-height: calc(1.5 * ${stepToRem(-2)});
+      font-size: var(--backingCard-description-fontSize);
+      line-height: calc(
+        (var(--backingCard-description-fontSize) + ${pxToRem(4)})
+      );
     }
   }
 
@@ -171,6 +197,34 @@ export const StyledBackingCard = styled.article`
     gap: ${pxToRem(10)};
   }
 
+  &.k-BackingCard--hasHover:hover:not(.k-BackingCard--disabled),
+  button&:hover:not(.k-BackingCard--disabled),
+  a&:hover:not(.k-BackingCard--disabled) {
+    --backingCard--border-color: var(--color-grey-500);
+
+    cursor: pointer;
+
+    .k-BackingCard__imageWrapper {
+      img,
+      figure,
+      video {
+        transform: scale(1.05);
+      }
+    }
+    .k-BackingCard__button:not(.k-Button--disabled) {
+      border-color: ${COLORS.primary2};
+      background-color: ${COLORS.primary2};
+    }
+    .k-BackingCard__description__moreButton {
+      color: ${COLORS.primary2};
+    }
+  }
+
+  button&:active,
+  a&:active {
+    --backingCard--border-color: var(--color-grey-600);
+  }
+
   &.k-BackingCard--disabled {
     color: var(--color-grey-500);
     cursor: not-allowed;
@@ -183,7 +237,8 @@ export const StyledBackingCard = styled.article`
       }
     }
 
-    .k-BackingCard__amount {
+    .k-BackingCard__amount,
+    .k-BackingCard__amount__suffix {
       color: var(--color-grey-500);
     }
 
@@ -200,33 +255,5 @@ export const StyledBackingCard = styled.article`
     .k-BackingCard__description__moreButton.k-u-color-primary1 {
       color: var(--color-grey-600) !important;
     }
-  }
-
-  &.k-BackingCard--hasHover:hover,
-  button&:hover,
-  a&:hover {
-    --backingCard--border-color: var(--color-grey-500);
-
-    cursor: pointer;
-
-    .k-BackingCard__imageWrapper {
-      img,
-      figure,
-      video {
-        transform: scale(1.05);
-      }
-    }
-    .k-BackingCard__button {
-      border-color: ${COLORS.primary2};
-      background-color: ${COLORS.primary2};
-    }
-    .k-BackingCard__description__moreButton {
-      color: ${COLORS.primary2};
-    }
-  }
-
-  button&:active,
-  a&:active {
-    --backingCard--border-color: var(--color-grey-600);
   }
 `